chore: Add new SI prefixes to size printer
The 27th General Conference on Weights and Measures, the governing body
behind SI prefixes, accepted in its resolution the inclusion of new SI
prefixes R ("ronna") and Q ("quetta") for 10^27 and 10^30, respectively.
Adding the associated binary prefixes, RiB and QiB (robibytes and
quebibytes) to the formatting logic.

It is highly unlikely that we will ever need to deal with these prefixes
in real use, nontheless.