feat(server/config): Implement config manager and decouple `worker_processes`
This commit introduces a new `ServerConfiguration` handler object which
duty is to deal with all the handling of the configuration options the
servers are started with. Previously, this was done by `SessionManager`
since a5119f04ea21dd6dda927d962fa1bc892131db9f introduced
non-authentication-related configuration options into the previously
authentication-only configuration file.

The new infrastructure aims to offer a generally user-friendly and more
type-safe way of dealing with configuration options in various parts of
the `server` package. Options are registered simply akin to adding
options to an `argparse` parser.

Moved the logic of finding or creating an example server config to the
new class.

The first option, `worker_processes` has already been refactored into
being handled by the new instance, but there are more to follow...