fix(migration): Prevent statement timeouts when upgrading `reports`
In case of a lot of reports, about 5M or more, and a small enough
`statement_timeout` (<5 minutes) setting in the case of PostgreSQL, the
large single query that previously dealt with assigning
`report.checker_id` (the new `INT` one) from `checkers` could time out.

This patch fixes this issue.