feat(test): Error if server fails to start for many minutes
If the `server` subprocess fails to start during testing, the only
symptom is that the tests hang indefinitely, and it is non-trivial to
change the configuration s.t. the error message of the server appears,
**especially** in the CI environment.

Instead, print the output file to the stdout of the spawner process (so
`pytest` captures it) and raise an explicit Exception to make sure that
the test suite appears failing "immediately".