feat(server): Multiprocess migration and `db_cleanup`
This patch allows the server to run the schema migration and
`db_cleanup` actions in parallel across the configured products, to
speed up these operations. Previously, the migration and cleanup ran in
a sequential job on all products, forfeiting the benefit of having
multiple CPUs on a server. As each product is a separate database and
there must not be any shared resource between products, it is safe to
run each migration in a separate process, in parallel.

Migrations and cleanup is prepared for scheduling in a deterministic
order, `ORDER BY endpoint`. (Previously it was done by the `ROWID`.)

The connection to the "config" database is released early on to
prevent a timeout on the unused and not changing configuration
database from crashing the server during a longer-running product
migration.

Added a facility to create beautiful logging output in migration
scripts and for the cleanup routines. This log output will now
necessarily include the product's `endpoint` identifier, as the log
messages are no longer sequential.